+++ src/gort/overwatcher/helpers/dome.py
@@ -173,9 +173,7 @@ class DomeHelper:
                 await self.stop()
                 status = await self.status()
 
-            await self.overwatcher.notify("Opening the dome ...")
             await self._move(status, open=True, park=park)
-            await self.overwatcher.notify("The dome is now open.")
 
     async def close(self, park: bool = True, retry: bool = False):
         """Closes the dome."""
@@ -195,14 +193,12 @@ class DomeHelper:
                 await self.stop()
                 status = await self.status()
 
-            await self.overwatcher.notify("Closing the dome ...")
             await self._move(
                 status,
                 open=False,
                 park=park,
                 retry_without_parking=retry,
             )
-            await self.overwatcher.notify("The dome is now closed.")
 
     async def stop(self):
         """Stops the dome."""